+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.63] - 2025-11-22
+### Added
+- `extraNumOfObservations` variable in the `Location` object
+
+### Changed
+- updated `Location` object docs
+
 ## [0.9.62] - 2024-11-22
 ### Added
 - support for 'Geocaching HQ Block Party' cache type